Get Wasm tests working
