+++ compress_test.go
@@ -1,10 +1,10 @@
+// +build !js
+
 package websocket
 
 import (
 	"strings"
 	"testing"
-
-	"cdr.dev/slog/sloggers/slogtest/assert"
 
 	"nhooyr.io/websocket/internal/test/xrand"
 )
@@ -15,14 +15,21 @@ func Test_slidingWindow(t *testing.T) {
 	const testCount = 99
 	const maxWindow = 99999
 	for i := 0; i < testCount; i++ {
-		input := xrand.String(maxWindow)
-		windowLength := xrand.Int(maxWindow)
-		r := newSlidingWindow(windowLength)
-		r.write([]byte(input))
+		t.Run("", func(t *testing.T) {
+			t.Parallel()
 
-		if cap(r.buf) != windowLength {
-			t.Fatalf("sliding window length changed somehow: %q and windowLength %d", input, windowLength)
-		}
-		assert.True(t, "hasSuffix", strings.HasSuffix(input, string(r.buf)))
+			input := xrand.String(maxWindow)
+			windowLength := xrand.Int(maxWindow)
+			r := newSlidingWindow(windowLength)
+			r.write([]byte(input))
+
+			if cap(r.buf) != windowLength {
+				t.Fatalf("sliding window length changed somehow: %q and windowLength %d", input, windowLength)
+			}
+
+			if !strings.HasSuffix(input, string(r.buf)) {
+				t.Fatalf("r.buf is not a suffix of input: %q and %q", input, r.buf)
+			}
+		})
 	}
 }
