+++ conn_test.go
@@ -4,32 +4,22 @@ package websocket_test
 
 import (
 	"context"
-	"io"
+	"fmt"
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"os/exec"
+	"sync"
 	"testing"
 	"time"
 
 	"golang.org/x/xerrors"
 
 	"nhooyr.io/websocket"
-	"nhooyr.io/websocket/internal/test/cmp"
 	"nhooyr.io/websocket/internal/test/wstest"
 	"nhooyr.io/websocket/internal/test/xrand"
+	"nhooyr.io/websocket/internal/xsync"
 )
-
-func goFn(fn func() error) chan error {
-	errs := make(chan error)
-	go func() {
-		defer func() {
-			r := recover()
-			if r != nil {
-				errs <- xerrors.Errorf("panic in gofn: %v", r)
-			}
-		}()
-		errs <- fn()
-	}()
-
-	return errs
-}
 
 func TestConn(t *testing.T) {
 	t.Parallel()
@@ -44,7 +34,7 @@ func TestConn(t *testing.T) {
 				ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
 				defer cancel()
 
-				copts := websocket.CompressionOptions{
+				copts := &websocket.CompressionOptions{
 					Mode:      websocket.CompressionMode(xrand.Int(int(websocket.CompressionDisabled) + 1)),
 					Threshold: xrand.Int(9999),
 				}
@@ -60,8 +50,8 @@ func TestConn(t *testing.T) {
 				defer c1.Close(websocket.StatusInternalError, "")
 				defer c2.Close(websocket.StatusInternalError, "")
 
-				echoLoopErr := goFn(func() error {
-					err := echoLoop(ctx, c1)
+				echoLoopErr := xsync.Go(func() error {
+					err := wstest.EchoLoop(ctx, c1)
 					return assertCloseStatus(websocket.StatusNormalClosure, err)
 				})
 				defer func() {
@@ -72,38 +62,12 @@ func TestConn(t *testing.T) {
 				}()
 				defer cancel()
 
-				c2.SetReadLimit(1 << 30)
+				c2.SetReadLimit(131072)
 
 				for i := 0; i < 5; i++ {
-					n := xrand.Int(131_072)
-
-					msg := xrand.Bytes(n)
-
-					expType := websocket.MessageBinary
-					if xrand.Bool() {
-						expType = websocket.MessageText
-					}
-
-					writeErr := goFn(func() error {
-						return c2.Write(ctx, expType, msg)
-					})
-
-					actType, act, err := c2.Read(ctx)
+					err := wstest.Echo(ctx, c2, 131072)
 					if err != nil {
 						t.Fatal(err)
-					}
-
-					err = <-writeErr
-					if err != nil {
-						t.Fatal(err)
-					}
-
-					if expType != actType {
-						t.Fatalf("unexpected message typ (%v): %v", expType, actType)
-					}
-
-					if !cmp.Equal(msg, act) {
-						t.Fatalf("unexpected msg read: %v", cmp.Diff(msg, act))
 					}
 				}
 
@@ -111,6 +75,44 @@ func TestConn(t *testing.T) {
 			})
 		}
 	})
+}
+
+func TestWasm(t *testing.T) {
+	t.Parallel()
+
+	var wg sync.WaitGroup
+	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		wg.Add(1)
+		defer wg.Done()
+
+		c, err := websocket.Accept(w, r, &websocket.AcceptOptions{
+			Subprotocols:       []string{"echo"},
+			InsecureSkipVerify: true,
+		})
+		if err != nil {
+			t.Error(err)
+			return
+		}
+		defer c.Close(websocket.StatusInternalError, "")
+
+		err = wstest.EchoLoop(r.Context(), c)
+		if websocket.CloseStatus(err) != websocket.StatusNormalClosure {
+			t.Errorf("echoLoop: %v", err)
+		}
+	}))
+	defer wg.Wait()
+	defer s.Close()
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*20)
+	defer cancel()
+
+	cmd := exec.CommandContext(ctx, "go", "test", "-exec=wasmbrowsertest", "./...")
+	cmd.Env = append(os.Environ(), "GOOS=js", "GOARCH=wasm", fmt.Sprintf("WS_ECHO_SERVER_URL=%v", wstest.URL(s)))
+
+	b, err := cmd.CombinedOutput()
+	if err != nil {
+		t.Fatalf("wasm test binary failed: %v:\n%s", err, b)
+	}
 }
 
 func assertCloseStatus(exp websocket.StatusCode, err error) error {
@@ -121,39 +123,4 @@ func assertCloseStatus(exp websocket.StatusCode, err error) error {
 		return xerrors.Errorf("unexpected close status (%v):%v", exp, err)
 	}
 	return nil
-}
-
-// echoLoop echos every msg received from c until an error
-// occurs or the context expires.
-// The read limit is set to 1 << 30.
-func echoLoop(ctx context.Context, c *websocket.Conn) error {
-	defer c.Close(websocket.StatusInternalError, "")
-
-	c.SetReadLimit(1 << 30)
-
-	ctx, cancel := context.WithTimeout(ctx, time.Minute)
-	defer cancel()
-
-	b := make([]byte, 32<<10)
-	for {
-		typ, r, err := c.Reader(ctx)
-		if err != nil {
-			return err
-		}
-
-		w, err := c.Writer(ctx, typ)
-		if err != nil {
-			return err
-		}
-
-		_, err = io.CopyBuffer(w, r, b)
-		if err != nil {
-			return err
-		}
-
-		err = w.Close()
-		if err != nil {
-			return err
-		}
-	}
 }
