+++ internal/xsync/go_test.go
@@ -1,0 +1,20 @@
+package xsync
+
+import (
+	"testing"
+
+	"nhooyr.io/websocket/internal/test/cmp"
+)
+
+func TestGoRecover(t *testing.T) {
+	t.Parallel()
+
+	errs := Go(func() error {
+		panic("anmol")
+	})
+
+	err := <-errs
+	if !cmp.ErrorContains(err, "anmol") {
+		t.Fatalf("unexpected err: %v", err)
+	}
+}
