+++ ws_js_test.go
@@ -1,4 +1,4 @@
-package websocket
+package websocket_test
 
 import (
 	"context"
@@ -6,25 +6,43 @@ import (
 	"os"
 	"testing"
 	"time"
+
+	"nhooyr.io/websocket"
+	"nhooyr.io/websocket/internal/test/cmp"
+	"nhooyr.io/websocket/internal/test/wstest"
 )
 
-func TestEcho(t *testing.T) {
+func TestWasm(t *testing.T) {
 	t.Parallel()
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
 	defer cancel()
 
-	c, resp, err := Dial(ctx, os.Getenv("WS_ECHO_SERVER_URL"), &DialOptions{
+	c, resp, err := websocket.Dial(ctx, os.Getenv("WS_ECHO_SERVER_URL"), &websocket.DialOptions{
 		Subprotocols: []string{"echo"},
 	})
-	assert.Success(t, err)
-	defer c.Close(StatusInternalError, "")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer c.Close(websocket.StatusInternalError, "")
 
-	assertSubprotocol(t, c, "echo")
-	assert.Equalf(t, &http.Response{}, resp, "http.Response")
-	echoJSON(t, ctx, c, 1024)
-	assertEcho(t, ctx, c, MessageBinary, 1024)
+	if !cmp.Equal("echo", c.Subprotocol()) {
+		t.Fatalf("unexpected subprotocol: %v", cmp.Diff("echo", c.Subprotocol()))
+	}
+	if !cmp.Equal(http.StatusSwitchingProtocols, resp.StatusCode) {
+		t.Fatalf("unexpected status code: %v", cmp.Diff(http.StatusSwitchingProtocols, resp.StatusCode))
+	}
 
-	err = c.Close(StatusNormalClosure, "")
-	assert.Success(t, err)
+	c.SetReadLimit(65536)
+	for i := 0; i < 10; i++ {
+		err = wstest.Echo(ctx, c, 65536)
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+
+	err = c.Close(websocket.StatusNormalClosure, "")
+	if err != nil {
+		t.Fatal(err)
+	}
 }
